Add Gradle as a build system
You can now use

```sh
./gradlew run
```

or

```sh
./gradlew run --args="hostname.example.com"
```

to run the project, and

```sh
./gradlew test
```

to run the JUnit tests.

The reason for this change is that, previously, there wasn't a way to
run the JUnit tests short of installing and running JUnit 4 by hand.

As part of this change, I've removed:

1. The `run-client.sh` script, which didn't work well with Gradle. (It
   wanted to be run from the `src` directory, while `./gradlew` wanted
   to be run from the root directory.)
   Where you previously would have used `./run-client.sh`, instead start the
   file server manually and then use `./gradlew run`.

2. The Eclipse configuration files, which seemed to be out-of-date
   anyway. They were still set to Java 7, while other lab files (eg,
   the GitHub Actions workflows) are set to Java 11.