Implement the main loop in FileRetriever.java
This loop collects the packets into files. Then, when all of the files
are complete, it writes them to disk.

The actual I/O (which is handled by OutOfMoneyApiService and
FileWriterService) is stubbed out right now; only the business logic is
written at the moment.

(This commit also includes unit tests for FileRetriever, which should
currently pass.)